Refactoring and improving image loading
