+++ src/sections/ViewPhotoAlbum/ViewPhotoAlbum.css
@@ -39,18 +39,38 @@
 }
 
 .photos .photos-container {
-  display: grid; 
+  display: grid;
   grid-template-columns: repeat(4, 1fr);
-  gap: 1rem .5rem;
+  gap: 1rem 0.5rem;
 }
 
 .photos .photos-container .wrapper {
   width: 100%;
   height: 100%;
+  transition: transform 0.5s ease;
+  overflow: hidden;
+}
+
+.photos .photos-container .wrapper .location {
+  font-size: x-small;
+  color: #777;
+}
+
+.photos .photos-container .wrapper .notes {
+  font-size: small;
 }
 
 .photos .photos-container .wrapper img {
   object-fit: cover;
   height: 100%;
   width: 100%;
+  min-height: 100%;
+  min-width: 100%;
+  transition: all 0.5s ease;
+  opacity: 0;
+}
+
+.photos .photos-container .wrapper img.photo:hover {
+  transform: scale(1.02);
+  cursor: pointer;
 }
