+++ src/sections/ViewPhotoAlbum/types.ts
@@ -1,0 +1,20 @@
+export interface PhotoLinksManifest {
+  hosts: StorageHostsInfo;
+  files: Array<PhotoInfo>;
+}
+
+export interface StorageHostsInfo {
+  compressed: StorageHostInfo;
+  full: StorageHostInfo;
+}
+
+export interface StorageHostInfo {
+  domain: string;
+  path: string;
+}
+
+export interface PhotoInfo {
+  filename: string;
+  location: string;
+  notes: string;
+}
